Adds Slash Command to list Topics subscribed by a Mattermost chan… (#30)
* Bump Mattermost to 5.18 (#29)

* Add a Slach Command skeleton code to Plugin

This patch adds the skeleton code for a Slash command
for the AWS SNS plugin. The functionality of the command
will be implemented in next patches.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>

* Store and retrieve Topics subscribed by the channel

This patch stores the Topic subscribed by the configured
channel upon confirmation. And when the slash command
`/awssns listTopics` is called, retrives the Topic from
the KV store and presents it to the channel.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>

* Use a map instead of slice for better handling of items

Using a map allows us to manipulate individual items. For
example we can delete the item without having to iterate
over all the items. Also allows to uniquely add a new
element.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>

* Delete only the topic from struct instead of whole struct

We only need to delete the topic element from the struct
when unsubscription confirmation is received. So here instead
of deleting the entry in KV Store we just delete the element
from the KV Store and push it back to KV Store.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>

* Resolving some of the linting error messages

This patch only resolves some of the linting error messages
But many of them especially from plugins file is left as it is.

They will be resolved in future patches

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>

* Moves list topics logic to a function

Instead of implementing the logic within the switch case, this PR
moves it to a function and calls the function.

Also, format the text with a bullet points and use a prefix for
KVSet and KVGet.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>

Co-authored-by: Ben Schumacher <ben.schumacher@mattermost.com>